Simplify form saving and return HTTP status code
Change form saving to use HTTP POST so a form is send rather than all
of the components of a form. This makes adding fields to a form cleaner.
Change saving a form to return an HTTP status code because front end code
expects a form that is not saved successfully to not return HTTP status 200.